Implement lockfree parallel work construct:
Allows a batch of work to be completed by a set of parallel logic,
but no (implicit or explicit) locking is necessary to get the result
of the parallel computation (i.e. we don't need to poll on some
completion counter and no one thread that popped a work item but
got preempted can block the entire system).

It works because it allows any thread to steal any partially completed
work from another thread and complete it on that thread's behalf. Of
course there is the caveat that some job code may be run more than
once, which is why there is a constraint that the job function must
publish any side effects atomically, and only succeed on the first run.
This may sound unfeasible at first, but using a lockfree set for results,
or using a CAS to publish it can make this practically useful for a large
number of cases.