lockfree_deque: Fix unsafe reference counting of nodes
Namely, two issues are addressed. First, the reference count
of a node must be incremented before a CAS operation since, in
concurrent situations, the node can be deleted as soon as the
CAS succeeds. Second of all, PushLeft/PushRight routines must
temporarily bump the reference count of a new node so that it
is retained until the end of the subroutine call. The pushed
nodes may still be dequeued by concurrent threads but the node
should not be released in order to maintain all the invariants
of the function.